requirements.txt
kafka-python==2.0.2
requests==2.31.0
pyspark==3.4.0
streamlit==1.28.0
pandas==2.0.3
plotly==5.17.0
scikit-learn==1.3.0